fix(api): current volume calculation in hardware controller (#18528)
Closes AUTH-1952

# Overview

When calculating current pipette volumes following liquid handling
commands, the hardware controller does floating point math. This was
resulting in small python-specific floating point errors which were
essentially negligible in all practical sense but were causing `==`
comparisons to fail.

For the specific bug reported above, this floating point error was
making the pipette's `current_volume` to be a value that's 'very close
to' zero, but not zero, after doing a full dispense. This was preventing
the harware controller from properly executing a
`prepare_for_aspirate()` as it needs the current volume to be `0`.

So, this PR prevents the bug from happening by making sure that the
final pipette volume after a full dispense gets saved as `0uL`

## Risk assessment

Low- bug fix only