chore: use hatch as a build backend for python projects (PD 8.5.0) (#18538)
This is a version of #18377 that is targeted on chore_release-pd-8.5.0.
This was a bit of a pain because I accidentally clicked the button that
merges edge into that pr and the merge had to be unwound, so I'm only
like 85% confident everything came with it.

This is a long-time-coming change to switch to using
[hatch](https://github.com/pypa/hatch) (or more specifically,
[hatchling](https://github.com/pypa/hatch/tree/master/backend)) as a
[PEP-517 build backend](https://peps.python.org/pep-0517/) for building
our python packages.
## What? Why?

Python packaging has been in a bit of a churn for a while. The way it
used to work is you wrote a script, called setup.py for convention, that
would import utilities from the built in package distutils and later its
drop in replacement setuptools that provided a function called setup();
then you called setup() with a bunch of descriptive data, and that
function would build some kind of python distributable, like an sdist or
a wheel (or, back in the day, an egg).

People largely hated this because it meant that you would have to run
code to build a package, and that's gross, and also there was no way to
figure out how to build a package than writing the code - maybe that
setup.py had some other dependencies, you'd never know until you tried.

The fix for this was to define an official build system API. Python
build systems are now split into backends and frontends. Backends are
packages that provide a specific programmatic interface for building
packages; frontends are applications that transform user input into the
appropriate calls to that interface. These two packages are now commonly
configured through the use of a toml file called pyproject.toml with a
specific format.

This happened a while ago, and the python ecosystem moved on to
supporting it. Specifically, setuptools switched to focusing on that,
and doesn't want to deal with setup.py anymore. Features traditionally
provided by setuptools to make setup.py work have been getting
deprecated or outright broken. Most recently, that was editable-install
support. This is even more of a problem because setuptools is commonly
included as a transitive dependency without a version bound, so it can
just update on you and break stuff.

Switching to use hatchling for building our projects instead of setup.py
lets us avoid these breakages by using a more stable build system; it
lets us configure our projects using more modern configuration files,
that can expand into also configuring our build frontends; and it can
all be done without actually changing the way we interact with these
projects as devs, because pip is a PEP-517 build frontend, and that's
what pipenv uses to install stuff. So we don't have to switch out pipenv
(yet) and everything will Just Keep Working.
How?

First step is to switch all the python projects to use pyproject.tomls
instead of setup.py, requesting hatchling as a build backend. There's
two things that don't Just Work with this setup:

- We get our version from git, which isn't uncommon, but we switch what
tag we look at sometimes, which is uncommon. [This hatch
plugin](https://github.com/Opentrons/hatch-plugins/tree/main/hatch-vcs-tunable)
we wrote lets you override the settings passed to the git version bridge
plugin via environment variables
- We exactly coversion all our dependencies, and we don't want to edit
the files to do it. [This hatch
plugin](https://github.com/Opentrons/hatch-plugins/tree/main/hatch-dependency-coversion)
we wrote automatically overwrites the versions of specified dependencies
at build time

We also need to switch https://github.com/Opentrons/buildroot/pull/246
and https://github.com/Opentrons/oe-core/pull/196 to use this.

## What's left?

- system builds working
  - [x] on ot2 https://github.com/Opentrons/buildroot/pull/246
  - [x] on flex  https://github.com/Opentrons/oe-core/pull/196
- [x] the other python packages
- [x] verifying that this works all the way through a built robot
version
  - [x] on ot2, to boot and run
  - [x] on flex
- verifying that push works
  - [x] on ot2
  - [x] on flex
- [x] verifying that the interned build for offline analysis works
- [ ] verifying that a pypi push works
- [x] checking that everybody is ok with it