chore(app-shell): disable windows signing better
Setting forceCodeSigning=false means that builds won't fail if "the app
would not be signed", but I think that means "if you haven't specified
signing details", not "the shell call to sign the executables fails" -
if the latter happens, the build fails anyway. So if we're not signing,
just don't even specify the sign options.