feat(protocol-designer,step-generation): remove blowoutOffsetFromTop in moveLiquid steps
Blowouts now occur at defined locations and are no longer configurable by the user. If the blowout location is selected to be the destination well, the blowout occurs at the retract point. If the blowout location is selected to be the source well, teh blowout occurs at the top of the well.