fix(hardware-testing): old liquid class method in gravimetric script (#18560)
# Overview

Updates the gravimetric test script to use the new `get_liquid_class()`
function and makes the gravimetric script always use the latest API version.

## Risk assessment

None for production code. 
High for testing code if updating the api version is going to have a big
impact on testing. If not, then low-none risk.