feat(protocol-designer): filter liquid class options and show warning (#18559)
This PR wires up liquid class filtering based on form field selection on
page 1 of moveLiquid and mix forms. It also wires up the warning logic
for showing why some (if any) liquid classes are disabled based on those
selections and for prioritizing which warning should show if there are
multiple reasons for disabling.

Closes AUTH-1530