chore(app-shell): install python for each arch
Significant refactor of python install path. Now, instead of installing
python to a staging location and then relying on electron to pack it,
we'll pack it ourselves. This has to happen because electron's universal
app packer requires all files to be identical by-hash between
architectures, which python install trees won't be. In theory we could
use universal python installs, but the resource we use for portable
python distributions suitable for embedding doesn't make them available.

We also split up the python install mechanics from the build hooks that
call them because I've been vacillating on when to do the python
install, and this makes that a bit easier.