chore: build a universal app on osx (#18599)
When the apple aarch64 Mx processors and subsequent macbook pros that
used them came out, we never actually started building the app for arm -
osx had just-in-time binary emulation called rosetta 2 that was good
enough, or so we thought.

Now, since we've updated electron-builder to a version new enough to do
so, we can build a universal app that has both x64 and aarch64 binaries
stapled together to form "fat mach-os" so that each architecture can use
whichever form of the code is faster. And we can tell: rosetta 2 was not
good enough. Why did we think this. It's a js vm, every part of it is a
hot loop, why did we ship this running in rosetta, it's one million
times better with an arm build.

There is a downside to doing this, though: size. Universal apps are not
magic; they work by just putting _both_ the x64 and the arm64 code into
the app package, so if the executable payloads are X MB, the universal
version of the app is X*2 MB. This is bad enough with electron, since
its framework (the osx version of the core dlls for the application) is
200MB or so and thus will be 400MB when universal, but it would be much
much worse for python. That's also true because the python osx universal
build is not distributed by our python redistributable source - mostly
because universal apps made with universal compilation as opposed to
post-facto mach-o stapling have to target the same core framework
version, so it would support old hardware but not old hardware on old
software, which the folks running the redistributable project don't want
to do - so we would need to just have 2 full python installations. That
means two full numpys (100MB on their own), two full pandas, two full
interpreter frameworks... all told, this would make the size-on-disk of
the installed app on osx over a gig.

Instead, let's just only give the universal treatment to the hotloop:
the core electron framework that is executing all our js. Offline
analysis is rare and not critical; we'll keep packaging just the x64
interpreter for now, and hold out hope that
https://github.com/astral-sh/python-build-standalone/issues/140 gets
addressed.

## Testing
- [x] run the new universal app and let your mind boggle at how much
faster it is. Oh my god. I'll put links here once they're built.
-  make sure offline analysis still works
- [x] on osx :
https://builds.opentrons.com/app/Opentrons-v8.5.0-alpha.2-mac-b57460-osx-build-universal-app-build-both.dmg
- [x] on windows:
https://builds.opentrons.com/app/Opentrons-v8.5.0-alpha.2-win-b57460-osx-build-universal-app-build-both.exe
- [x] on linux:
https://builds.opentrons.com/app/Opentrons-v8.5.0-alpha.2-linux-b57460-osx-build-universal-app-build-both.AppImage

Unfortunately, the other OSs need to be handled too because we put
python in a slightly different place now