refactor(step-generation, app): generate get_liquid_class() from step… (#18612)
… forms

Upon talking with @ncdiehl11, i realized that the `get_liquid_class()`
should be based off of the liquid class selected in the step forms
itself other than the liquid class assigned to the liquid. this is
because the assigned liquid is the default and you can still change the
liquid class type in the step form. so the step form liquid class is
most accurate.