refactor(components): Allow opting out of <LabwareRender> self-positioning (#18590)
## Overview

Closes EXEC-1573.

## Changelog

* Add an optional prop to `<LabwareRender>` that allows a parent
component to take over responsibility for offsetting the labware origin
within a slot. See EXEC-1573 for background.

As of this PR, no parent components actually do this. I'm migrating them
incrementally in separate PRs.

  Eventually, I'm aiming for the old mode to be totally deleted.

* Same for `<Labware>`.

* Add some util functions to shared-data. These will be used to help
parent components compute the appropriate offset. They're not called
anywhere yet as of this PR, but I've been playing around with them in a
draft branch, so I'm pretty confident that they're basically the right
thing.

---------

Co-authored-by: Jamey Huffnagle <jamey.huffnagle@opentrons.com>