feat(step-generation): more compact Python generation for dicts (#18592)
# Overview

The custom liquid class dict for `define_liquid_class()` is kind of big.
This PR makes the dict formatting a bit more compact: if all the entries
in the dict are scalar, and their total length is short-ish, we will
emit the dict on a single line.

Before:
```
"offset": {
    "x": 1,
    "y": 0,
    "z": 5,
},
```
After:
```
"offset": {"x": 1, "y": 0, "z": 5},
```
(See tests for other examples.)

This PR also cleans up the code for emitting lists of pairs as Python
tuples that was introduced in PR #18558.

## Test Plan and Hands on Testing

Added/updated unit tests.

## Risk assessment

Low.