fix(protocol-designer): don't show errors until attempting to save/continue (#18627)
Fixes logic for showing form errors, especially when reopening an
errored-out saved form. We previously used a state variable for whether
or not to show the form errors, such that they would only show if a user
tried to continue/save a form containing an error. After a large error
handling [refactor](https://github.com/Opentrons/opentrons/pull/18437)
was merged, this state was unused, such that errors would show
automatically when reopening a form.

This PR adds logic to the `makeSingleEditFieldProps` util to show form
errors in the following scenarios:
1. the error is due to the field being required, and the form is saved
and reopened
2. the error is of any type, the form is pre-saved or saved, and the
user attempts to continue or save

I also wire up a new optional property of the form error to specify
whether or not to show the error upon reopening in scenario 1 above.