refactor(app,components,labware-library): Support labware-schema-3-style positioning in some components (#18557)
## Overview

Goes towards EXEC-1278.

## Details

To prepare for labware schema 3, we want to move math and logic out of
React components and into new utils in `shared-data`. Those utils can
then abstract over differences between labware schemas 2 and 3, and
implement whatever positioning features we add to labware schema 3.

By "math and logic," I mean things like: What do you choose as the SVG
`viewbox`? Within that `viewbox`, how do you position the SVG elements
comprising the labware? If you need to render a labware in a slot, on a
module, or stacked atop another labware, how does that affect things?

There are basically two cases. Each case already has a util in
`shared-data`, added in #18590.

1. We're rendering a labware on its own, like in the thumbnails in the
labware gallery.
2. We're rendering a labware relative to other things, like on a deck
map.

So, in this PR:

* We switch a bunch of components from manually constructing `viewBox`es
and `transform`s to delegating to the appropriate util from
`shared-data`.
* We set `<Labware positioningMode="passThrough" ... />`, which the new
utils assume. See EXEC-1573 for background.

This is the first of probably several PRs that will do this. The
selection of components so far is basically random.