refactor(robot-server, api): Remove `definition` from `loadModule` result (#18639)
Closes EXEC-212 and works toward EXEC-756 and EXEC-77

The loadModule command returns the module definition as a part of the command result. This isn't great, because module definitions are not customer facing, and they create versioning risks. We're removing definition from loadModule command now because locating features will touch module definitions, and not removing definitions means we will have to handle a quite complex database migration as an alternative.