fix(api): sanitize labware labels
Labware labels are passed through from the protocol api all the way to
the commands and eventually the client with little actual
_functionality_ dependent on their types or their values. That means
that the only thing that checks them is pydantic model validation, and
if you avoid that - as the legacy command mapper does by using
model_construct() - then other downstream things won't be happy.

In this case, OT-2 LPC starts its session by echoing commands it reads
from the robot analysis back out to a maintenance run. When a client
uses HTTP to create commands, pydantic model validation _does_ run, and
it would fail, and this wasn't an anticipated error scenario because,
well, it's echoed from the protocol! But the incorrect data made it
through.

Another way to handle this would be to do pydantic model validation on
the protocol inputs, but that would break any old protocols that
implicitly depended on this working. Compatibility! Also IMO it's silly
that we enforce this if the entire system works just fine if you give it
an integer.

Closes RESC-423