fix(api): Fix gripper move point calculation for modules (#18691)
Closes RQA-4295

When we get_labware_origin_position, we start by retrieving the deck coordinates for the underlying addressable area, so for example, if the addressable area is heaterShakerV1D1, we get the coordinates for D1. Much later down stream, if there is a module involved in a stackup, we offset D1 by the heaterShakerV1D1 's offsetFromCutoutFixture.

In #18564, we refactored get_labware_grip_point to utilize more of the get_labware_origin_position logic (see that PR for more details), however, the grip point does exactly one thing different: we utilize the deck position for the actual addressable area we care about, ex, heaterShakerV1D1. This means that an extra offsetFromCutoutFixture is applied for grip moves.

To fix, we should refactor get_labware_origin_position to get the deck coordinates for the addressable area we actually care about and not do the offsetFromCutoutFixture adjustment much later down stream. Doing so unifies the gripper point and labware point calculation behavior almost entirely.