fix(app): queue notification refetch if mqtt topic update occurs while refetching (#18694)
Closes RABR-781

If MQTT informs the app that it should refetch a particular topic while the app is actively refetching that topic, the app doesn't perform an additional refetch after the first refetch completes. This creates a race potential condition.

To fix, the app always needs to refetch again if MQTT tells the app an update is available while the app is in the process of refetching.