feat(api): add keep_last_tip argument to liquid class transfer methods (#18693)
Adds a new argument `keep_last_tip` to three liquid class transfer functions (`transfer_with_liquid_class`,
`consolidate_with_liquid_class`, and `distribute_with_liquid_class`). If this is set to `True`, the last (or only tip for `never` and `once`) will not be dropped at the end of the high level transfer/consolidate/distribute. Otherwise, the last tip will be dropped, even if the tip policy is `never`.