feat(api): do prepare for aspirate and airgap above well always (#18653)
Closes AUTH-1978
Follows #18585 

# Overview

When we aren't using LLD or estimating liquid heights based on well
volumes, we cannot verify if the retract position set in a liquid class
lies inside liquid. In such cases we run the risk of accidentally
aspirating liquid inside the tip (and eventually inside the pipette). So
this PR moves the pipette to a safe location above the well (2mm above
well top) before executing a `prepareForAspirate` or adding an air gap.

## Changelog

- added a check to compare the retract location and a safe location for
plunger movement (2mm above well top) and do a move to safe location if
needed
- added unit tests

## Risk assessment

Low. Supposed to reduce risk of accidental liquid aspiration.