feat(api): add new tip option of 'always' to consolidate and distribute (#18695)
Closes AUTH-1995

# Overview

Allows `new_tip="always"` as a valid tip option for LC-based consolidate
and distribute. This will make the new consolidate and distribute
functions behave same as PD:
- 'always' for consolidate will pick up a new tip before **the next
chunk** of aspirates. Or in other words, will pick up a new tip when
going back to aspirate *after* a dispense
- 'always' for distribute will pick up a new tip before every aspirate

*Note that this behavior is different from what 'always' does in the
old, non-LC-based consolidate and distribute*. The old functions would
pick up a tip only once in the beginning of the function even if the tip
option was 'always'.

## Risk assessment

Medium. Touches a part of existing code that determines tip pickups.