chore: refactor to cli application
This was running under pytest but it wasn't really _using_ pytest, and
therefore didn't have the upsides like test collection and
parametrization to balance the downsides like it being annoying to add
command line flags. By making it a standalone application, we get all
that stuff.

While we're at it, refactor so it'll be easier to add new kinds of tests
that take advantage of the test spec generation and so on.