chore(api): Speed up stackup snapshot tests (#18700)
@SyntaxColoring and @mjhuff added these extremely sick snapshot tests
for labware geometry. Only downside was they took forever, but they're
awesome and a solid basis for confidently modifying and improving that
logic in the future. By parallelizing them - which was really easy since
they were already doing subprocess stuff, so there was no shared logic
to break - we can cut the test execution time down to something
reasonable to run in your inner loop.

Also, there's not much of a point to this being in pytest since we're
not taking advantage of anything it can do while everything's in one big
test (which it has to be to take advantage of the multiprocessing); so
pop it out into a new little integration-testing place and make it a
command-line application that can now do filtering and such.

## to come out of draft
- [x] test that the filtering stuff works as well as it ought to
- [x] fix the log and stdout/stderr spam from the console
- [ ] do something about running out of file descriptors lol