fix(api): require LC transfers to use API v2.24 (#18712)
Closes AUTH-2004

# Overview

Updates the liquid classes based transfer, consolidate and distribute
methods to require 2.24 instead of 2.23. Details in ticket above for why
we need to do this.

## Risk assessment

Not really. We had previously indicated in api source code that liquid
classes work is in-progress and subject to change. And these methods are
not available in public docs until 8.5.0.