feat(step-generation): ensure retract position is safe for air gap (#18730)
This PR introduces logic to check whether or not the retract position is
safe for post-aspirate or -dispense air gaps. If the retract position is
2mm or more above the top of the well (deemed to be a "safe" position
for air gap), we can move to the retract position for air gap.
Otherwise, we move to the safe position above the well, and air gap
there.