fix(api): error raising when there's no innergeometry for a well (#18728)
Closes RQA-4304

# Overview

The check that verifies that a submerge start location or retract end
location is not inside the liquid in well relies on knowing the height
of the liquid- either by calculating it based on liquid volume and well
inner geometry or by probing into the well.
We already skip the check if we don't know the liquid volume or haven't
probed the well yet, opting to simply log the decision instead. We
should also skip the check if a well's innergeometry isn't present
either otherwise the check will raise errors for any labware that
doesn't have inner geometry defined, like the `smc_384_read_plate`, or
if it's a custom labware.

So this PR makes it so that the check excepts on
`IncompleteLabwareDefinitionError` as well.

## Risk assessment

Low. Only skips a check that would have errored without being useful.