fix(protocol-designer): use liquid class defaults when resetting tip position
This PR adds logic to pull the correct default tip position given the tip position type (aspirate, dispense, aspirate_retract, aspirate_submerge, dispense_retract, dispense_submerge). Previously, the reset behavior was hard coded to set the tip 1mm from bottom of the well. This is accurate for aspirate and dispense, but not for retract and submerge.

Closes RQA-4256