chore(ci): include matrix.with-ot-hardware in test name to tell tests apart (#18745)
# Overview

Some time ago, we added `with-ot-hardware: ['true', 'false']` to the
test matrix for the `opentrons package tests`, but we didn't update the
test name with the new variable. This makes it look like the same test
is running twice in CI, which is confusing.

This change just adds the `with-ot-hardware` to the test name.

## Test Plan and Hands on Testing

Run the CI tests and see.

## Risk assessment

Low, cosmetic change for Github display.