fix(protocol-designer): migrate null blowout location
Previously, we erroniously allowed a user to save a form with disposal or blowout checkbox on, but no blowout location selected from the dropdown. For the 8.5.0 release, we enforce a stepform error if the blowout location is left unselected. In order to prevent this error popping up for users who migrate older protocols in which the blowout location was unselected, we migrate the blowout location to the first available trash or waste chute ID.

Closes RQA-4311