fix(shared-data): revert changes to LC displayNames (#18774)
# Overview

In PR #18711 (AUTH-2003), we changed the liquid class display names to:
`Aqueous (Deionized water)` / `Volatile (80% ethanol)` / `Viscous (50% glycerol)`.

But we've decided we want to change the display names back. So:
- displayName: `Aqueous`, description `Deionized water`, loadname: `water`
- displayName: `Volatile`, description `80% ethanol`, loadname: `ethanol_80`
- displayName: `Viscous`, description `50% glycerol`, loadname: `glycerol_50`

https://opentrons.slack.com/archives/C07JJA1AJFQ/p1751036532344239?thread_ts=1750692306.637929&cid=C07JJA1AJFQ

This PR reverts #18711 and #18738, and was generated by `git revert`
with no manual changes:
```
git revert ec54caf 968ddc8
```

## Test Plan and Hands on Testing

Running CI tests.

## Review requests

This is what we want now, right?

## Risk assessment

Low, but we're cutting it close.