Code refactor changes
This removes a lot of HTML and replaces that with markdown. Sticking to a more pure markdown build. A lot of the HTML for images is not needed and can be done better with markdown.

Added CSS to manage image sizes and caption styles.

Fixed some table headers. Tags were <td> when they should be <th>.

Fixed WEEE text and in-line image. Added float-right image.