fix(protocol-designer): only show liquid class assignment if in source well
This PR fixes the subtext in the liquid class options in a transfer or mix form. We should only show "Assigned to [x,y,z wells]" subtext to a liquid class option if the liquid class is assigned to a liquid in at least one of the selected source/mix wells. If the liquid class is present in the source/mix labwrae, but not present in the selected source wells, we show the liquid class's generic description.

Closes RQA-4106