fix(api): don't overwrite `next_dest`, fix distribute double-dispense (#18802)
# Overview

In `distribute_with_liquid_class()`, we were dispensing into the same
destination well twice when the distribute consists of multiple
tip-fills.

The bug was caused by the reuse of the variable `next_dest` in
`distribute_with_liquid_class()`. `next_dest` in the outer loop is
supposed to contain the next well that we have not dispensed to yet.
However, we accidently overwrote the variable because we also use it as
a loop variable for an unrelated inner loop:
```
for next_vol, next_dest in vol_dest_combo:
```

## Test Plan and Hands on Testing

Using `analyze`, I confirmed that the double-dispense happened with
Andy's example before this change, and that it was fixed after this
change.

I also updated
`test_order_of_water_distribution_steps_using_multi_dispense()` to
perform a distribute that requires multiple tip-fills, and changed the
test to check that we're dispensing into specific wells instead of
`dest=mock.ANY`. I confirmed that the bug happened before this change,
and it was fixed after this change.

## Risk assessment

Medium: this is change deep inside our implementation. But it's
necessary to fix the bug.