feat(protocol-designer): add warning about dangerous submerge/retract position (#18825)
In analysis, we determine whether the submerge/retract location for a
given liquid transfer is deemed to be within the specified well's
liquid. We are afforded this since the API calculates the liquid
meniscus height in a all wells throughout the protocol. However, we do
not have access to this logic in Protocol Designer. Due to this reality,
we cannot know for sure whether a submerge/retract position set inside a
well is safely above the meniscus level. The resolution is to warn the
user that their submerge/retract position may result in analysis failure
if the tip is in fact below the calculated liquid meniscus at the time
of the pipette movement.

Closes RQA-4334