fix(app): stacker error recovery shouldn't call `manualRetrieve` pre-maturely (#18824)
This PR fixes the following error recovery issues:
1. We skipped a step in the `STACKER_SHUTTLE_EMPTY_SKIP` route by
routing to the wrong step in `HoldingLabware`
2. The `flexStackerLabwareRetrieveFailed` error was not caught in our
failedLabwareUtil so we've been rendering incorrect labware info for the
`STACKER_SHUTTLE_EMPTY_SKIP/RETRY` routes
3. The re-engage latch PNG was not rendered properly because we rendered
it as a video instead of an image
4. `ReleaseLabware` screen now renders a latch release animation instead
of gripper for stacker-related flows
5. We prematurely called `manualRetrieve` and changed the stacker state
before exiting the flow. When a user tries to go back to a previous
step, the rest of the flow would always end up failing. To fix this, the
`manualRetrieve` happens in the last step of the flow.
6. Updating the hopper labware info screen to reflect the correct number
after a potential manual retrieval, since now we won't get the state
change to occur until the flow is completed